refactor: minor corrections
