+++ src/lib/locationState/stores/types.ts
@@ -7,7 +7,7 @@ export type Store = {
 
   set(name: string, value: unknown): void;
 
-  load(key?: string): void;
+  load(key: string): void;
 
   save(): void;
 };
